src/third_party/blink/web_tests/virtual/web-components-v0-disabled/shadow-dom/v0/v0-added-after-v1-and-slotted-expected.txt

CONSOLE ERROR: line 22: Uncaught TypeError: v0host.createShadowRoot is not a function
